requirements.txt
Flask>=3.0,<4.0
Flask-Login>=0.6,<1.0
Flask-WTF>=1.2,<2.0
Flask-Mail>=0.10,<1.0
Flask-Limiter>=3.5,<4.0
bcrypt>=4.1,<5.0
cryptography>=42.0,<44.0
openai>=1.12,<2.0
python-magic>=0.4,<1.0
Jinja2>=3.1,<4.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
